formatica/requirements.txt
streamlit 
opencv-python 
pytesseract 
pillow
streamlit-cropper
